requirements.txt
pandas~=2.3.0
scikit-learn~=1.7.0
numpy~=2.3.1
pytest~=8.4.1
matplotlib~=3.10.3
seaborn~=0.13.2
mlflow-skinny~=3.1.1
fastapi~=0.115.14
pydantic~=2.11.7